# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Monitoring Platform API",
    "description": "API for system monitoring, alerts, notifications, and reporting",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/alerts/secureSave": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "summary": "Save alert definition (secure)",
        "description": "Save alert with path validation",
        "operationId": "saveAlertSecure",
        "parameters": [
          {
            "name": "alertName",
            "in": "query",
            "description": "Alert name",
            "required": false,
            "schema": {
              "type": "string",
              "default": "custom-alert"
            }
          },
          {
            "name": "content",
            "in": "query",
            "description": "Alert content",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name: Custom Alert\ntype: threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/alerts/save": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "summary": "Save alert definition",
        "description": "Create or update alert definition file",
        "operationId": "saveAlert",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "description": "Alert filename",
            "required": false,
            "schema": {
              "type": "string",
              "default": "alert-custom.yml"
            }
          },
          {
            "name": "content",
            "in": "query",
            "description": "Alert content",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name: Custom Alert\ntype: threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/alerts/secureContent": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get alert definition (secure)",
        "description": "Read alert with path validation",
        "operationId": "getAlertSecure",
        "parameters": [
          {
            "name": "alertName",
            "in": "query",
            "description": "Alert name",
            "required": false,
            "schema": {
              "type": "string",
              "default": "cpu-high"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/alerts/content": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "Get alert definition",
        "description": "Read alert definition file",
        "operationId": "getAlert",
        "parameters": [
          {
            "name": "filename",
            "in": "query",
            "description": "Alert filename",
            "required": false,
            "schema": {
              "type": "string",
              "default": "alert-cpu-high.yml"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}